Factor out api functionality
