+++ app/client/components/identity/NewsletterSection.tsx
@@ -1,8 +1,8 @@
 import React, { FC } from "react";
 import palette from "../../colours";
 import { DropMenu } from "./DropMenu";
-import { ConsentOption, Theme } from "./identity";
 import { MarketingPreference } from "./MarketingPreference";
+import { ConsentOption, Theme } from "./models";
 import { PageSection } from "./PageSection";
 
 type ClickHandler = (id: string) => {};
